add user id to post

+++ backend/routes/posts.js
@@ -43,9 +43,9 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imgPath: url + "/images/" + req.file.filename
+      imgPath: url + "/images/" + req.file.filename,
+      owner: req.userData.userId
     });
-    console.log(post)
     post.save().then(createdPost => {
       res.status(201).json({
         message: "Post added successfully",
@@ -72,11 +72,15 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imgPath: imgPath
+      imgPath: imgPath,
+      owner: req.userData.userId
     });
-    console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
+    Post.updateOne({ _id: req.params.id, owner: req.userData.userId }, post).then(result => {
+      if (result.nModiied > 0) {
+        res.status(200).json({ message: "Update successful!" });
+      } else {
+        res.status(401).json({ message: "Unauthorized User" });
+      }
     });
   }
 );
@@ -119,10 +123,19 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
+  Post.deleteOne({ _id: req.params.id, owner: req.userData.userId }).then(result => {
     console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
-  });
+      if (result.n > 0) {
+        res.status(200).json({ message: "Deletion successful!" });
+      } else {
+        res.status(401).json({ message: "Not authorized!" });
+      }
+    })
+    .catch(error => {
+      res.status(500).json({
+        message: "Deleting posts failed!"
+      });
+    });
 });
 
 module.exports = router;
